requirements.txt
py-cord==2.3.2
numpy==1.22.3
opencv_python==4.6.0.66
Pillow==9.3.0
